# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Rentway API'
  description: 'Mobile-friendly REST API v1. Secured endpoints require a Bearer token obtained via /api/v1/auth/login. API Owner: Sudharma HewaVitharana.'
  version: 1.0.0
servers:
  -
    url: /api
    description: 'API base path'
paths:
...
  /v1/invoices:
    get:
      tags:
        - Invoices
      summary: 'List invoices'
      operationId: 1640e2df2c8c25690fabfa827fe64d50
      parameters:
        -
          name: company_id
          in: query
          required: false
          schema:
            type: integer
        -
          name: invoice_status
          in: query
          required: false
          schema:
            type: string
        -
          name: customer_id
          in: query
          required: false
          schema:
            type: integer
        -
          name: per_page
          in: query
          required: false
          schema:
            type: integer
      responses:
        200:
          description: OK
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Invoices
      summary: 'Create invoice'
      operationId: 45849a810ca4e57e72044ec023de2deb
      responses:
        201:
          description: Created
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        422:
          $ref: '#/components/responses/ValidationError'
      security:
        -
          bearerAuth: []
  '/v1/invoices/{id}':
    get:
      tags:
        - Invoices
      summary: 'Show invoice'
      operationId: 9f2522103fb3b115721dc943988f2c3b
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
      security:
        -
          bearerAuth: []
  /v1/notifications:
    get:
      tags:
        - Notifications
      summary: 'List notifications'
      operationId: 9e3dfdc4faf54aa01f2e07fda5374455
      parameters:
        -
          name: company_id
          in: query
          required: false
          schema:
            type: integer
        -
          name: type
          in: query
          required: false
          schema:
            type: string
        -
          name: read_state
          in: query
          required: false
          schema:
            type: string
        -
          name: per_page
          in: query
          required: false
          schema:
            type: integer
      responses:
        200:
          description: OK
      security:
        -
          bearerAuth: []
  '/v1/notifications/{id}':
    get:
      tags:
        - Notifications
      summary: 'Show notification'
      operationId: 9043744c2caba58e1e8cc72bdedf9e0e
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Notifications
      summary: 'Delete notification'
      operationId: 3214f17d13ba232374a55aa5b0d077e6
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Deleted
      security:
        -
          bearerAuth: []
...
  /v1/reports/summary:
    get:
      tags:
        - Reports
      summary: 'Financial summary'
      operationId: a30b97bc29e5351198a39377f199bd68
      parameters:
        -
          name: company_id
          in: query
          required: false
          schema:
            type: integer
        -
          name: month
          in: query
          required: false
          schema:
            type: string
        -
          name: year
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
      security:
        -
          bearerAuth: []
...
  /v1/vehicles:
    get:
      tags:
        - Vehicles
      summary: 'List vehicles'
      operationId: eb6e2c396cab1218870893536e8325b4
      responses:
        200:
          description: OK
    post:
      tags:
        - Vehicles
      summary: 'Create vehicle'
      operationId: 679667a7bea11c54470817ecd9d32761
      responses:
        201:
          description: Created
      security:
        -
          bearerAuth: []
  '/v1/vehicle/{id}':
    get:
      tags:
        - Vehicles
      summary: 'Show vehicle'
      operationId: bd1ffda92f66c586b7d9ef40feafa0f2
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
  '/v1/vehicles/{id}':
    put:
      tags:
        - Vehicles
      summary: 'Update vehicle'
      operationId: ac3001b3499ed25b7cb09a1f57bd6125
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Updated
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Vehicles
      summary: 'Delete vehicle'
      operationId: 2abc417e7b23df1cc31a6b970327f963
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Deleted
      security:
        -
          bearerAuth: []
components:
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Unauthorized
            type: object
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Forbidden
            type: object
    ValidationError:
      description: 'Validation failed'
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 'Validation failed'
              errors:
                type: object
            type: object
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...